+++ src/main/java/org/codinjutsu/tools/jenkins/logic/AuthenticationNotifier.java
@@ -12,5 +12,5 @@ public interface AuthenticationNotifier {
 
     void loginCancelled();
 
-    void loginFailed(Exception ex);
+    void loginFailed(Throwable ex);
 }