+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## [Unreleased]
-- \#309  Exception in plugin Jenkins Control
+- \#309 Exception in plugin Jenkins Control
+- \#322 Nullpointer if Jenkins if not running
 
 ## [0.13.12]
 - \#280 render DateParameterDefinition as textfield